requirements.txt
psycopg2-binary>=2.9.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
sqlalchemy>=1.4.0
